+++ sdk/rolldown.config.js
@@ -88,6 +88,7 @@ export default defineConfig([
             'physics3d/index': 'src/physics3d/index.ts',
             'spine/index': 'src/spine/index.ts',
             'tilemap/index': 'src/tilemap/index.ts',
+            'logic/index': 'src/logic/index.ts',
             'dragonbones/index': 'src/dragonbones/index.ts',
             'douyin/index': 'src/platform/douyin/index.ts',
             'wasm': 'src/wasm.ts',