+++ sdk/rollup.config.js
@@ -19,6 +19,7 @@ export default [
             'douyin/index': 'src/platform/douyin/index.ts',
             'spine/index': 'src/spine/index.ts',
             'tilemap/index': 'src/tilemap/index.ts',
+            'logic/index': 'src/logic/index.ts',
             'dragonbones/index': 'src/dragonbones/index.ts',
             'wasm': 'src/wasm.ts',
         },